+++ aqt/archives.py
@@ -20,7 +20,10 @@
 # IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 # CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-from six.moves import urllib
+try:
+    from urllib.request import ProxyHandler, build_opener, install_opener, urlopen
+except ImportError:
+    from urllib2 import ProxyHandler, build_opener, install_opener, urlopen
 import xml.etree.ElementTree as ElementTree
 
 
@@ -62,10 +65,10 @@ class QtArchives:
 
         # Get packages index
         update_xml_url = "{0}Updates.xml".format(archive_url)
-        proxies = urllib.request.ProxyHandler({})
-        opener = urllib.request.build_opener(proxies)
-        urllib.request.install_opener(opener)
-        content = urllib.request.urlopen(update_xml_url).read()
+        proxies = ProxyHandler({})
+        opener = build_opener(proxies)
+        install_opener(opener)
+        content = urlopen(update_xml_url).read()
         self.update_xml = ElementTree.fromstring(content)
         for packageupdate in self.update_xml.iter("PackageUpdate"):
             name = packageupdate.find("Name").text