+++ CHANGELOG.rst
@@ -19,9 +19,13 @@ Added
 Changed
 -------
 
+* Project URL is changed.
+
 Fixed
 -----
 
+* Update README wordings.
+* Remove dependency for python3 with aqtinst command utility.
 * Fix command name in help message.
 
 Deprecated